+++ CHANGELOG.md
@@ -4,6 +4,12 @@
 
 ### Added
 
+### Changed
+
+## 0.2.2
+
+### Added
+
 - Allow for headers with OME-TIFF.
 - Add COVID-19 Image to public demo.
 - Added `getMetadata` function to `ZarrLoader` and `OMETiffLoader` classes to enable Vitessce to obtain human-readable image layer metadata.